+++ demo/uni-app/components/jyf-parser/libs/MpHtmlParser.js
@@ -1,7 +1,7 @@
 /**
  * html 解析器
  * @tutorial https://github.com/jin-yufeng/Parser
- * @version 20200513
+ * @version 20200521
  * @author JinYufeng
  * @listens MIT
  */
@@ -127,6 +127,7 @@ class MpHtmlParser {
 				name: 'img',
 				attrs: {
 					src: 'data:image/svg+xml;utf8,' + src.replace(/#/g, '%23'),
+					style: (/vertical[^;]+/.exec(node.attrs.style) || []).shift(),
 					ignore: 'T'
 				}
 			})
@@ -168,15 +169,11 @@ class MpHtmlParser {
 		if ((this.options.compress & 2) && attrs.class) attrs.class = void 0;
 		switch (node.name) {
 			case 'a':
-			case 'ad':
-				this.bubble();
-				break;
-				// #ifdef APP-PLUS
+			case 'ad': // #ifdef APP-PLUS
 			case 'iframe':
-			case 'embed':
+				// #endif
 				this.bubble();
 				break;
-				// #endif
 			case 'font':
 				if (attrs.color) {
 					styleObj['color'] = attrs.color;
@@ -195,6 +192,24 @@ class MpHtmlParser {
 					attrs.size = void 0;
 				}
 				break;
+			case 'embed':
+				// #ifndef APP-PLUS
+				var src = node.attrs.src || '',
+					type = node.attrs.type || '';
+				if (type.includes('video') || src.includes('.mp4') || src.includes('.3gp') || src.includes('.m3u8'))
+					node.name = 'video';
+				else if (type.includes('audio') || src.includes('.m4a') || src.includes('.wav') || src.includes('.mp3') || src.includes(
+						'.aac'))
+					node.name = 'audio';
+				else break;
+				if (node.attrs.autostart)
+					node.attrs.autoplay = 'T';
+				node.attrs.controls = 'T';
+				// #endif
+				// #ifdef APP-PLUS
+				this.bubble();
+				break;
+				// #endif
 			case 'video':
 			case 'audio':
 				if (!attrs.id) attrs.id = node.name + (++this[`${node.name}Num`]);
@@ -212,9 +227,11 @@ class MpHtmlParser {
 					}
 				}
 				attrs.source = [];
-				if (attrs.src) attrs.source.push(attrs.src);
-				if (!attrs.controls && !attrs.autoplay)
-					console.warn(`存在没有 controls 属性的 ${node.name} 标签，可能导致无法播放`, node);
+				if (attrs.src) {
+					attrs.source.push(attrs.src);
+					attrs.src = void 0;
+				}
+				if (!attrs.controls && !attrs.autoplay) attrs.controls = 'T';
 				this.bubble();
 				break;
 			case 'td':
@@ -517,7 +534,7 @@ class MpHtmlParser {
 				if (this.STACK[i].name == name) break;
 			if (i != -1) {
 				var node;
-				while ((node = this.STACK.pop()).name != name);
+				while ((node = this.STACK.pop()).name != name) this.popNode(node);
 				this.popNode(node);
 			} else if (name == 'p' || name == 'br')
 				this.siblings().push({